melhorando componetização e alterações da iteração2
